Move capsule from experimental to v1
Refer to:
https://review.openstack.org/543878
