+++ acceptance/openstack/container/v1/capsules_test.go
@@ -1,17 +1,17 @@
-package experimental
+package v1
 
 import (
 	"testing"
 
 	"github.com/gophercloud/gophercloud/acceptance/clients"
-	"github.com/gophercloud/gophercloud/openstack/container/experimental/capsules"
+	"github.com/gophercloud/gophercloud/openstack/container/v1/capsules"
 	th "github.com/gophercloud/gophercloud/testhelper"
 )
 
 func TestCapsuleGet(t *testing.T) {
-	client, err := clients.NewContainerExperimentalClient()
+	client, err := clients.NewContainerV1Client()
 	if err != nil {
-		t.Fatalf("Unable to create an container experimental client: %v", err)
+		t.Fatalf("Unable to create an container v1 client: %v", err)
 	}
 	th.AssertNoErr(t, err)
 	capsuleUUID := "e6c913bb-b4e4-409d-8b71-3e029f196458"
