+++ openstack/client.go
@@ -409,7 +409,7 @@ func NewMessagingV2(client *gophercloud.ProviderClient, clientID string, eo goph
 	return sc, err
 }
 
-// NewContainerExperimental creates a ServiceClient that may be used with experimental container package
-func NewContainerExperimental(client *gophercloud.ProviderClient, eo gophercloud.EndpointOpts) (*gophercloud.ServiceClient, error) {
-	return initClientOpts(client, eo, "container-experimental")
+// NewContainerV1 creates a ServiceClient that may be used with v1 container package
+func NewContainerV1(client *gophercloud.ProviderClient, eo gophercloud.EndpointOpts) (*gophercloud.ServiceClient, error) {
+	return initClientOpts(client, eo, "container")
 }
