+++ openstack/container/v1/capsules/testing/requests_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/gophercloud/gophercloud"
-	"github.com/gophercloud/gophercloud/openstack/container/experimental/capsules"
+	"github.com/gophercloud/gophercloud/openstack/container/v1/capsules"
 	th "github.com/gophercloud/gophercloud/testhelper"
 	fakeclient "github.com/gophercloud/gophercloud/testhelper/client"
 )
