+++ openstack/container/v1/capsules/urls.go
